Moved srs to special and fixed tests
