+++ sunpy/io/tests/test_srs.py
@@ -1,23 +1,23 @@
 """
 This module implements tests for SRS Reader.
 """
-import sunpy.io.srs, pytest
-from sunpy.io.srs import read
-
-import sunpy.data.test
 import os
 
+import pytest
+
+from sunpy.io.special import srs
+import sunpy.data.test
 
 testpath = sunpy.data.test.rootdir
 
-srs_20150906 = os.path.join(testpath, '20150906SRS.txt')
-srs_20150306 = os.path.join(testpath, '20150306SRS.txt')
-srs_20150101 = os.path.join(testpath, '20150101SRS.txt')
-@pytest.mark.paramterize("path_, number_of_rows",
-                    [(srs_20150906, 5),
-                     (srs_20150306, 4),
-                     (srs_20150101, 9)])
-def test_number_of_rows(path_, number_of_rows):
-    table = sunpy.io.srs.read(path_)
+filenames = [{'file': '20150906SRS.txt', 'rows': 5},
+             {'file': '20150306SRS.txt', 'rows': 4},
+             {'file': '20150101SRS.txt', 'rows': 9}]
+
+@pytest.mark.parametrize("path, number_of_rows",
+                         [(os.path.join(testpath, elem['file']), elem['rows'])
+                          for elem in filenames])
+def test_number_of_rows(path, number_of_rows):
+    table = srs.read(path)
     assert len(table) == number_of_rows
-    
+
